Create finding positions of stars
I've plotting the positions of the two variable stars that we have data on - RZ Ceph and RV UMa. What you can do it change the date of observation and find their position (alt and az) using this program as well as some plots which I think are quite useful to show their movement throughout the sky. I've been speaking to someone else doing out project and they suggest picking stars that are close to these two stars so I'll try get that done for tomorrow or do it in the session.